1.0.0

+++ src/main/java/com/origins_eternity/sanity/content/capability/sanity/ISanity.java
@@ -29,7 +29,7 @@ public interface ISanity {
 
     void setDizzy(boolean dizzy);
 
-    void consumeSanity(float value);
+    void consumeSanity(double value);
 
-    void recoverSanity(float value);
+    void recoverSanity(double value);
 }
